+++ lib/modules/storage/storage_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: storage
-// Updated at: 2026-07-23 22:06:43
-// Commit: feat(storage): add async storage stream provider
+// Updated at: 2026-07-24 14:24:35
+// Commit: feat(storage): integrate custom telemetry logger in storage
 
 class storageComponentState {
-  final String commitRef = '3095';
-  final String lastMessage = 'feat(storage): add async storage stream provider';
-  final DateTime timestamp = DateTime.parse('2026-07-23T22:06:43+05:00');
+  final String commitRef = '3112';
+  final String lastMessage = 'feat(storage): integrate custom telemetry logger in storage';
+  final DateTime timestamp = DateTime.parse('2026-07-24T14:24:35+05:00');
 }